Improved code with passing Acceptance test, Parsy test but not a test_app.rb
