+++ lib/notes.html
@@ -1,0 +1,38 @@
+<!DOCTYPE html>
+<html>
+<head>
+	<title></title>
+</head>
+<body style="background-color: #f7f7f7">
+<table style="width:100%">
+<th>The Notes:</th>
+  <tr>
+    <td>Add 1 to 2</td>
+    <td>1 + 2  # => 3</td>
+  </tr>
+  <tr>
+    <td>Subtract 5 from 2 </td> 
+    <td>2 - 5  # => -3</td> 
+  </tr>
+  <tr>
+    <td>Is 1 less than 2</td>
+    <td>1 < 2  # => true</td>
+  </tr>
+  <tr>
+	<td>Is 1 equal to 2</td> 
+	<td>1 == 2 # => 3</td>
+  </tr>
+</table>
+<br>
+   <fieldset>
+      <form action="/search" method="get">
+        <label>
+          Search
+        </label>
+        <input type="text" name="query">
+        <br>
+   		<input type="submit" style="float: right;">
+      </form>
+    </fieldset>
+</body>
+</html>
